REFACTOR: Changed names and location of components for better clarity and structure
