+++ app/projects/components/ProjectsView.tsx
@@ -16,13 +16,13 @@ import React, { useState } from "react";
 import { AiOutlineClear } from "react-icons/ai";
 import { BsFilterLeft } from "react-icons/bs";
 import { ArchiveToggle } from "../../../components/Filters/ArchiveToggle";
-import ProjectsListSection from "./ProjectListSection";
+import ProjectsList from "../../../components/MaterialLists/ProjectsList";
 
 type ProjectsListProps = {
   allProjects: ProjectInterface[];
 };
 
-const ProjectsList: React.FC<ProjectsListProps> = ({ allProjects }) => {
+const ProjectsView: React.FC<ProjectsListProps> = ({ allProjects }) => {
   //^ Hooks
   const [isFilterOpen, setIsFilterOpen] = useState(false);
   const searchParams = useSearchParams();
@@ -448,7 +448,7 @@ const ProjectsList: React.FC<ProjectsListProps> = ({ allProjects }) => {
       />
 
       {/* List of projects */}
-      <ProjectsListSection groupedProjects={groupedProjects} />
+      <ProjectsList groupedProjects={groupedProjects} />
 
       {/* Filter Modal */}
       <FilterOverlay
@@ -466,4 +466,4 @@ const ProjectsList: React.FC<ProjectsListProps> = ({ allProjects }) => {
   );
 };
 
-export default ProjectsList;
+export default ProjectsView;
